Moved JWT token to infisical and moved SecretLoader class to Service namespace
