+++ FoodplannerApi/appsettings.json
@@ -7,7 +7,6 @@
   },
   "AllowedHosts": "*",
   "ApplicationSettings": {
-    "JWT_Secret": "Foodplanner_secret_giraf_multi_projcet_102930572926483932629376594y629",
     "JWT_Issuer": "Aalborg University",
     "JWT_Audience": "Egebakken",
     "JwtExpireDays": 30
