+++ FoodplannerServices/Auth/AuthService.cs
@@ -6,6 +6,8 @@ using System.Text;
 using FoodplannerModels.Account;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
+using Infisical.Sdk;
+using FoodplannerServices.Secret;
 
 namespace FoodplannerApi.Helpers
 {
@@ -27,7 +29,7 @@ namespace FoodplannerApi.Helpers
                 new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString())
             };
 
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["ApplicationSettings:JWT_Secret"]));
+            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(SecretsLoader.GetSecret("JWT_SECRET")));
             var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
 
             var token = new JwtSecurityToken(
